feat: refactor API routes into individual modules for better organization and maintainability
