+++ server/src/types/schemas.ts
@@ -13,7 +13,8 @@ export const SearchSchema = z.object({
   input: z.string().min(1, '"input" cannot be empty'),
   model: z.string(),
   session_id: z.string().min(1, '"session_id" cannot be empty'),
-  space_id: z.string(),
+  space_ids: z.array(z.string()).optional(),
+  file_ids: z.array(z.string()).optional(),
   provider: ProviderSchema,
   thread_id: z.string(),
 });
